make regions reuseable
extract dictionaries to utils.py
